Start included release
I'm separating the included release from the main release entity. This
is partially to avoid a circular reference in the entity, which would
deserialize just fine, but also because the data that is included in
the included entities isn't the full amount of data that is in the main
retrieved entity. If someone needs all of the data, they can get it
eventually by just querying for the component directly.